backend/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.0
python-dotenv==1.0.0
eventlet==0.33.0  # Recommended for SocketIO asynchronous operations
